+++ apps/web/src/app/playlist/edit/[playlist_id]/page.tsx
@@ -1,6 +1,6 @@
 import { notFound } from 'next/navigation';
 
-import { buildNoindexMetadata } from '../../../../../lib/seo/buildNoindexMetadata';
+import { buildNoindexMetadata } from '../../../../lib/seo/buildNoindexMetadata';
 import { getSSRAuthService } from '../../../../utils/auth/ssrAuth';
 import { PlaylistEditPageClient } from './PlaylistEditPageClient';
